+++ design-system/comment/src/styled/FieldStyles.ts
@@ -1,4 +1,5 @@
-import styled, { css } from 'styled-components';
+import { css } from '@emotion/core';
+import styled from '@emotion/styled';
 
 import { N500 } from '@atlaskit/theme/colors';