+++ design-system/select/docs/docs-util/propStatus.tsx
@@ -1,5 +1,5 @@
 import React, { Component, Fragment, MouseEvent, ReactNode } from 'react';
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 import Modal, { ModalTransition, OnCloseHandler } from '@atlaskit/modal-dialog';
 import Button from '@atlaskit/button/custom-theme-button';