+++ design-system/badge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/badge
 
+## 14.2.0
+
+### Minor Changes
+
+- [`f6b951a51f2`](https://bitbucket.org/atlassian/atlassian-frontend/commits/f6b951a51f2) - Removes usage of styled-components in favour of standardising on emotion
+
 ## 14.1.1
 
 ### Patch Changes