+++ design-system/datetime-picker/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/datetime-picker
 
+## 10.3.0
+
+### Minor Changes
+
+- [`f6b951a51f2`](https://bitbucket.org/atlassian/atlassian-frontend/commits/f6b951a51f2) - Removes usage of styled-components in favour of standardising on emotion
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.2.1
 
 ### Patch Changes